Move conversion function to separate file
